refactor(utilities): move rich text conversion utilities
