+++ package.json
@@ -21,6 +21,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
     "joi": "^17.12.2",
+    "jsdom": "^24.0.0",
     "lodash": "^4.17.21",
     "payload": "^2.11.1",
     "react": "^18.2.0",
@@ -46,6 +47,7 @@
     "@types/deep-equal": "^1.0.4",
     "@types/express": "^4.17.21",
     "@types/jest": "^29.5.12",
+    "@types/jsdom": "^21.1.6",
     "@types/node": "20.11.19",
     "@types/uuid": "^9.0.8",
     "@typescript-eslint/eslint-plugin": "6.21.0",
@@ -79,4 +81,3 @@
     "node": "16.20.0"
   }
 }
-
