+++ plugin/src/lib/utilities/tests/convertSlateToHtml.spec.ts
@@ -1,4 +1,4 @@
-import { convertSlateToHtml } from "..";
+import { convertSlateToHtml } from "../richTextConversion";
 import { payloadSlateToHtmlConfig } from "@slate-serializers/html";
 
 describe("convertSlateToHtml", () => {
